Configured dbt in Airflow
