+++ airflow/config/.dbt/README
@@ -1,0 +1,3 @@
+This is a folder created uniquely for enabling the use of dbt in Airflow. 
+It is a copy of the dbt profiles.yml file I use in local, adapted to the Airflow container.
+It is therefore NOT intended to be used locally.
